+++ environment/interpretation/lane_detector_test/src/targetFromLane/GetLaneBinary.jpg
