+++ environment/interpretation/lane_detector_test/src/targetFromLane/road_image.png
